AArch64: Use long for pointers in graphics/shader classes
Long is used to store native pointers in graphics/Shader
classes to support 64-bit architecture.

In addition, classes (Paint and GLES20Canvas) which directly
access native pointers of shader classes have been changed.

Note that other native pointers of Paint and GLES20Canvas
classes have not been changed to long. This will be done
as part of a separate patch.

+++ core/java/android/view/GLES20Canvas.java
@@ -1466,7 +1466,7 @@ class GLES20Canvas extends HardwareCanvas {
         return MODIFIER_NONE;
     }
 
-    private static native void nSetupShader(int renderer, int shader);
+    private static native void nSetupShader(long renderer, long shader);
     private static native void nSetupColorFilter(int renderer, int colorFilter);
     private static native void nSetupShadow(int renderer, float radius,
             float dx, float dy, int color);
